Added constraits for guestusers

+++ app/Http/Controllers/UserController.php
@@ -14,7 +14,11 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::withTrashed()->get();
+        $whitelist = array('id', 'name', 'picture', 'role', 'deleted_at');
+        if (auth('sanctum')->user()) {
+            array_push($whitelist, 'email', 'phone_number');
+        } 
+        return User::withTrashed()->get($whitelist);
     }
 
 
@@ -55,6 +59,10 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $authenticatedUser = auth('sanctum')->user();
+        if ($authenticatedUser && $authenticatedUser->id === $id) {
+            return response(['message'=>'You cant delete yourself!'] ,404);
+        }
         return User::find($id)->delete();
     }
